+++ src/accelerate/utils/__init__.py
@@ -107,6 +107,8 @@ from .imports import (
     is_tensorboard_available,
     is_timm_available,
     is_torch_xla_available,
+    is_torchdata_available,
+    is_torchdata_stateful_dataloader_available,
     is_torchvision_available,
     is_transformer_engine_available,
     is_transformers_available,
