+++ tests/test_accelerator.py
@@ -35,9 +35,8 @@ from accelerate.test_utils.testing import (
     require_non_torch_xla,
     require_torchdata_stateful_dataloader,
 )
-from accelerate.utils import patch_environment
+from accelerate.utils import patch_environment, is_torchdata_stateful_dataloader_available
 from accelerate.utils.dataclasses import DataLoaderConfiguration
-from accelerate.utils.imports import is_torchdata_stateful_dataloader_available
 from accelerate.utils.modeling import get_state_dict_from_offload, load_checkpoint_in_model
 from accelerate.utils.random import set_seed
